add docs and rename util classes.
