+++ example/flink/src/main/java/org/apache/iotdb/flink/FlinkTsFileBatchSource.java
@@ -34,11 +34,14 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+/**
+ * The example of reading TsFile via Flink DataSet API.
+ */
 public class FlinkTsFileBatchSource {
 
 	public static void main(String[] args) throws Exception {
 		String path = "test.tsfile";
-		Utils.writeTsFile(path);
+		TsFlieUtils.writeTsFile(path);
 		new File(path).deleteOnExit();
 		String[] filedNames = {
 			QueryConstant.RESERVED_TIME,
