+++ flink-tsfile/src/main/java/org/apache/iotdb/flink/tsfile/util/TSFileConfigUtil.java
@@ -22,6 +22,9 @@ package org.apache.iotdb.flink.tsfile.util;
 import org.apache.iotdb.tsfile.common.conf.TSFileConfig;
 import org.apache.iotdb.tsfile.common.conf.TSFileDescriptor;
 
+/**
+ * Utils of setting global TSFileConfig.
+ */
 public class TSFileConfigUtil {
 	
 	public static void setGlobalTSFileConfig(TSFileConfig config) {
